moving tests to new package structure
