+++ src/test/java/com/novoda/notils/java/string/StringUtilsTest.java
@@ -1,6 +1,4 @@
-package com.novoda.notils.string;
-
-import com.novoda.notils.java.string.StringUtils;
+package com.novoda.notils.java.string;
 
 import java.util.ArrayList;
 import java.util.Collection;
